LPS-7925 Upgrade
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@46653 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/upgrade/v6_0_0/util/DLFileEntryVersionUpgradeColumnImpl.java
@@ -48,9 +48,9 @@ public class DLFileEntryVersionUpgradeColumnImpl extends BaseUpgradeColumnImpl {
 	}
 
 	public Object getNewValue(Object oldValue) throws Exception {
-		double version = (Double)oldValue.doubleValue();
+		Double version = (Double)oldValue;
 
-		return getNewVersion(version);
+		return getNewVersion(version.doubleValue());
 	}
 
 }
